New function file/ Cleaner code with wrapper
